Add Answer Entity
