+++ src/main/java/com/ll/medium/domain/article/answer/Answer.java
@@ -1,5 +1,6 @@
-package com.ll.medium.domain.article.entity;
+package com.ll.medium.domain.article.answer;
 
+import com.ll.medium.domain.article.article.entity.Article;
 import com.ll.medium.domain.member.entity.Member;
 import jakarta.persistence.*;
 import lombok.Getter;
@@ -10,23 +11,18 @@ import java.time.LocalDateTime;
 @Entity
 @Getter
 @Setter
-public class Article {
+public class Answer {
     @Id
     @GeneratedValue(strategy= GenerationType.IDENTITY)
     private Long id;
 
-    private String title;
-
-    private String body;
+    private String content;
 
     private LocalDateTime createDate;
 
-    private boolean isPublished;
+    @ManyToOne
+    private Article article;
 
     @ManyToOne
     private Member author;
-
-    private LocalDateTime modifyDate;
-
-    private Integer hit;
 }
