+++ src/main/java/com/ll/medium/domain/home/home/controller/HomeController.java
@@ -1,7 +1,7 @@
 package com.ll.medium.domain.home.home.controller;
 
-import com.ll.medium.domain.article.entity.Article;
-import com.ll.medium.domain.article.service.ArticleService;
+import com.ll.medium.domain.article.article.entity.Article;
+import com.ll.medium.domain.article.article.service.ArticleService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Page;
 import org.springframework.stereotype.Controller;
