+++ src/main/resources/templates/domain/home/home_list.html
@@ -5,7 +5,6 @@
         <div class="text-end me-3">
             <a th:href="@{/post/write}" class="btn btn-primary">글쓰기</a>
         </div>
-
         <table class="table my-1">
             <thead>
             <tr class="text-center">
@@ -20,7 +19,8 @@
             <tr th:each="article, loop : ${paging}">
                 <td class="text-center align-middle" th:text="${loop.count}"></td>
                 <td class="align-middle">
-                    <a class="link-underline-light" th:href="@{|/post/${article.id}/increaseHit|}" th:text="${article.title}"></a>
+                    <a class="link-underline-light" th:href="@{|/post/${article.id}/increaseHit|}"
+                       th:text="${article.title}"></a>
                 </td>
                 <td>
                     <div th:if="${article.author != null}">
